completed assign experiments

+++ controllers/teachersClassController.js
@@ -1,24 +1,61 @@
-const LabSetup = require('../models/labSetup')
-const { Question } = require('../models/question')
-const { Teacher } = require('../models/teacher')
-const { TeacherClass } = require('../models/teacherClass')
+const LabSetup = require("../models/labSetup");
+const { Question } = require("../models/question");
+const { Teacher } = require("../models/teacher");
+const { Student } = require("../models/student");
+const { TeacherClass } = require("../models/teacherClass");
+
+async function inviteStudent(req, res) {
+    const { student_email } = req.body;
+    try {
+        const teacher = await Teacher.findOne({ _id: req.teacher._id });
+
+        const isTeacher = await Teacher.findOne({ email: student_email });
+
+        if (isTeacher) return res.status(400).send({ message: "eamil exist as teacher" });
+
+        const isStudent = await Student.findOne({ email: student_email });
+
+        const teacherStudents = teacher.students;
+
+        if (isStudent) {
+            const isTeacherStudent = teacherStudents.find((data) => data.student == isStudent._id);
+
+            if (isTeacherStudent) {
+                return res.status(404).send({ message: "can't add students that are not your student to a class" });
+            }
+        }
+
+        if (!isStudent) {
+            // create student
+            // send email notification
+        }
+
+        await teacher.save();
+        await TeacherClass.deleteOne({ _id: classId });
+        res.status(204).send(true);
+    } catch (error) {
+        if (error.kind === "ObjectId") return res.status(404).send({ message: "Class Not found" });
+
+        res.status(500).send({ message: "something went wrong" });
+        console.log(error.message);
+    }
+}
 
 async function deleteUnpublishedClass(req, res) {
-    const { classId } = req.params
+    const { classId } = req.params;
     try {
-        let teacher = await Teacher.findOne({ _id: req.teacher._id })
+        let teacher = await Teacher.findOne({ _id: req.teacher._id });
 
-        teacher = teacher.deleteClassById(classId)
+        teacher = teacher.deleteClassById(classId);
 
-        await teacher.save()
-        await TeacherClass.deleteOne({ _id: classId })
-        res.status(204).send(true)
+        await teacher.save();
+        await TeacherClass.deleteOne({ _id: classId });
+        res.status(204).send(true);
     } catch (error) {
-        if (error.kind === 'ObjectId')
-            return res.status(404).send({ message: 'Class Not found' })
+        if (error.kind === "ObjectId") return res.status(404).send({ message: "Class Not found" });
 
-        res.status(500).send({ message: 'something went wrong' })
-        console.log(error.message)
+        res.status(500).send({ message: "something went wrong" });
+        console.log(error.message);
     }
 }
 
@@ -28,49 +65,46 @@ async function getStudents(req, res) {
                 _id: req.params.classId,
             })
             .populate({
-                path: 'students',
-                select: 'name email imageUrl avatar _id isAccepted',
+                path: "students",
+                select: "name email imageUrl avatar _id isAccepted",
             })
-            .select('students, teacher')
+            .select("students, teacher");
 
-        if (!classData) return res.status(404).send({ message: 'Class not found' })
+        if (!classData) return res.status(404).send({ message: "Class not found" });
 
         if (classData.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
+            return res.status(401).send({ message: "Not autorized!" });
 
-        res.send(classData)
+        res.send(classData);
     } catch (error) {
-        res.status(500).send({ message: 'something went wrong' })
-        console.log(error)
+        res.status(500).send({ message: "something went wrong" });
+        console.log(error);
     }
 }
 
 async function addStudentToClass(req, res) {
-    const { studentId } = req.body
+    const { studentId } = req.body;
     try {
         let teacherClass = await TeacherClass.findOne({
             _id: req.params.classId,
-        })
+        });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         if (teacherClass.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
+            return res.status(401).send({ message: "Not autorized!" });
 
-        const isStudent = teacherClass.checkStudentById(studentId)
-        if (isStudent)
-            return res.status(400).send({ message: 'Student already added to class' })
+        const isStudent = teacherClass.checkStudentById(studentId);
+        if (isStudent) return res.status(400).send({ message: "Student already added to class" });
 
-        teacherClass = teacherClass.addStudentToClass(studentId)
-        await teacherClass.save()
+        teacherClass = teacherClass.addStudentToClass(studentId);
+        await teacherClass.save();
 
-        res.send(true)
+        res.send(true);
     } catch (error) {
-        res.status(500).send({ message: 'something went wrong' })
-        if (error.kind === 'ObjectId')
-            return res.status(404).send({ message: 'Class not found' })
-        console.log(error.message)
+        res.status(500).send({ message: "something went wrong" });
+        if (error.kind === "ObjectId") return res.status(404).send({ message: "Class not found" });
+        console.log(error.message);
     }
 }
 
@@ -78,153 +112,143 @@ async function getAllQuiz(req, res) {
     try {
         const teacherClass = await TeacherClass.findOne({
             _id: req.params.classId,
-        }).populate({ path: 'classwork.quiz' })
+        }).populate({ path: "classwork.quiz" });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         if (teacherClass.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
+            return res.status(401).send({ message: "Not autorized!" });
 
-        res.send(teacherClass)
+        res.send(teacherClass);
     } catch (error) {
-        res.status(400).send({ message: 'Invalid ID' })
-        console.log(error.message)
+        res.status(400).send({ message: "Invalid ID" });
+        console.log(error.message);
     }
 }
 async function getAllLab(req, res) {
     try {
         const teacherClass = await TeacherClass.findOne({
             _id: req.params.classId,
-        })
+        });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         if (teacherClass.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
+            return res.status(401).send({ message: "Not autorized!" });
 
-        res.send(teacherClass.classwork.lab)
+        res.send(teacherClass.classwork.lab);
     } catch (error) {
-        res.status(400).send({ message: 'Invalid ID' })
-        console.log(error.message)
+        res.status(400).send({ message: "Invalid ID" });
+        console.log(error.message);
     }
 }
 
 async function deleteQuiz(req, res) {
-    const questionId = req.params.questionId
+    const questionId = req.params.questionId;
     try {
         const teacherClass = await TeacherClass.findOne({
             _id: req.params.classId,
-        })
+        });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         if (teacherClass.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
-        const quiz = teacherClass.classwork.quiz
+            return res.status(401).send({ message: "Not autorized!" });
+        const quiz = teacherClass.classwork.quiz;
 
-        const index = quiz.findIndex((q) => q.toString() === questionId.toString())
-        if (index < 0)
-            return res.status(404).send({ message: 'Question Not found' })
+        const index = quiz.findIndex((q) => q.toString() === questionId.toString());
+        if (index < 0) return res.status(404).send({ message: "Question Not found" });
 
-        quiz.splice(index, 1)
-        await Question.deleteOne({ _id: questionId })
-        await teacherClass.save()
+        quiz.splice(index, 1);
+        await Question.deleteOne({ _id: questionId });
+        await teacherClass.save();
 
-        res.status(204).send(true)
+        res.status(204).send(true);
     } catch (error) {
-        res.status(400).send({ message: 'Invalid ID' })
-        console.log(error.message)
+        res.status(400).send({ message: "Invalid ID" });
+        console.log(error.message);
     }
 }
 async function deleteLab(req, res) {
-    const labId = req.params.labId
+    const labId = req.params.labId;
     try {
         const teacherClass = await TeacherClass.findOne({
             _id: req.params.classId,
-        })
+        });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         if (teacherClass.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
-        const lab = teacherClass.deleteLabById(labId)
+            return res.status(401).send({ message: "Not autorized!" });
+        const lab = teacherClass.deleteLabById(labId);
 
-        if (!lab) return res.status(400).send({ message: 'Lab not found' })
+        if (!lab) return res.status(400).send({ message: "Lab not found" });
 
-        await LabSetup.deleteOne({ _id: labId })
-        await teacherClass.save()
+        await LabSetup.deleteOne({ _id: labId });
+        await teacherClass.save();
 
-        res.status(204).send(true)
+        res.status(204).send(true);
     } catch (error) {
-        res.status(400).send({ message: 'Invalid ID' })
-        console.log(error.message)
+        res.status(400).send({ message: "Invalid ID" });
+        console.log(error.message);
     }
 }
 
 async function getClass(req, res) {
     try {
-        const teacherClass = await TeacherClass.findOne({ _id: req.params.classId })
+        const teacherClass = await TeacherClass.findOne({ _id: req.params.classId });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         // if (teacherClass.teacher.toString() !== req.teacher._id.toString())
         //     return res.status(401).send({ message: 'Not autorized!' })
 
-        res.send(teacherClass)
+        res.send(teacherClass);
     } catch (error) {
-        if (error.kind === 'ObjectId')
-            return res.status(404).send({ message: 'Class Not found' })
+        if (error.kind === "ObjectId") return res.status(404).send({ message: "Class Not found" });
 
-        res.status(500).send({ message: 'something went wrong' })
-        console.log(error.message)
+        res.status(500).send({ message: "something went wrong" });
+        console.log(error.message);
     }
 }
 
 async function deleteStudentFromClass(req, res) {
-    const { classId, studentId } = req.params
+    const { classId, studentId } = req.params;
 
     try {
-        let teacherClass = await TeacherClass.findOne({ _id: classId })
+        let teacherClass = await TeacherClass.findOne({ _id: classId });
 
-        if (!teacherClass)
-            return res.status(404).send({ message: 'Class not found' })
+        if (!teacherClass) return res.status(404).send({ message: "Class not found" });
 
         if (teacherClass.teacher.toString() !== req.teacher._id.toString())
-            return res.status(401).send({ message: 'Not autorized!' })
+            return res.status(401).send({ message: "Not autorized!" });
 
-        if (!teacherClass.removeStudentFromClass(studentId))
-            return res.status(404).send({ message: 'student not found' })
+        if (!teacherClass.removeStudentFromClass(studentId)) return res.status(404).send({ message: "student not found" });
 
-        teacherClass.removeStudentFromClass(studentId)
+        teacherClass.removeStudentFromClass(studentId);
 
-        await teacherClass.save()
-        res.status(204).send(true)
+        await teacherClass.save();
+        res.status(204).send(true);
     } catch (error) {
-        console.log(error.message)
-        if (error.kind === 'ObjectId')
-            return res.status(400).send({ message: 'Invalid class Id' })
-        res.status(500).send({ message: 'Something went wrong' })
+        console.log(error.message);
+        if (error.kind === "ObjectId") return res.status(400).send({ message: "Invalid class Id" });
+        res.status(500).send({ message: "Something went wrong" });
     }
 }
 
 async function getPublishedClassData(req, res) {
-    const { classId } = req.params
-    if (!classId) return res.status(400).send({ message: 'class not found' })
+    const { classId } = req.params;
+    if (!classId) return res.status(400).send({ message: "class not found" });
 
     try {
         const quizs = await TeacherClass.findOne({ _id: classId }).populate({
-            path: 'sentQuiz sentLab',
-            select: '-teacher',
-        })
-        res.send(quizs)
+            path: "sentQuiz sentLab",
+            select: "-teacher",
+        });
+        res.send(quizs);
     } catch (error) {
-        console.log(error)
-        res.status(500).send({ message: 'Something went wrong' })
+        console.log(error);
+        res.status(500).send({ message: "Something went wrong" });
     }
 }
 module.exports = {
@@ -238,4 +262,4 @@ module.exports = {
     getPublishedClassData,
     getClass,
     getStudents,
-}
+};
